requirements.txt
streamlit
openai
pandas
sqlalchemy
python-dot-env
